pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic[email]>=2.11.3",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "pydantic-extra-types>=2.10.3",
    "uvicorn>=0.34.0",
    "sqlmodel>=0.0.24",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.4.0",
    "aiosmtplib>=4.0.0",
    "phonenumbers>=9.0.3",
    "alembic>=1.15.2",
    "python-multipart>=0.0.20",
    "fastapi-offline>=1.7.3",
    "requests>=2.32.3",
]
